Licenta/back/requirements.txt
quart==0.20.0
quart-jwt-extended==0.1.0
quart-cors==0.7.0
aiohttp==3.9.5
aiosmtplib==2.0.0
redis==5.0.1
python-dotenv==1.0.1
firebase-admin==6.5.0
python-engineio
aioodbc==0.5.0
bcrypt==4.2.0
python-socketio
websockets
uvicorn